Moved APIKeySetup and fixed imports
